chatbot_function_calling/requirements.txt
langchain
python-dotenv
openai==0.28
uvicorn
fastapi